server/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
web3==7.6.0
python-dotenv==1.0.1
django-redis==5.4.0
redis==5.2.1